+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v2.0.0-rc.5
  * css-components: Refine style of list item title and subtitle.
  * ons-toolbar: Fix [#1419](https://github.com/OnsenUI/OnsenUI/issues/1419).
  * css-components: Refine default paddings of list items.
+ * ons-back-button: Add `onClick` property.
 
 v2.0.0-rc.4
 ----